Update the list of configuration variables
The only way to tell the default behavior of the extension was to look in the code.

This commit changes the structure of the doc section related to config variables to be presented as a table and to list default values. Some of the missing config variables were added along the way.